+++ Code/View/ZakaziKontrolu.xaml.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using Model.Appointment;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
@@ -20,20 +21,56 @@ namespace health_clinicClassDiagram.View
     /// </summary>
     public partial class ZakaziKontrolu : UserControl
     {
-        public ZakaziKontrolu()
+        private Appointment appointment;
+
+        public Appointment Appointment { get => appointment; set => appointment = value; }
+
+        public ZakaziKontrolu(Appointment appointment)
         {
+            Appointment = appointment;
             InitializeComponent();
         }
 
         private void homeButton_Click(object sender, RoutedEventArgs e)
         {
-            int thisCount = (this.Parent as Panel).Children.IndexOf(this);
-            (this.Parent as Panel).Children.RemoveRange(3, thisCount);
+            String message = "Ako napustite pregled sve izmene koje ste napravili će biti poništene\n\nDa li ste sigurni da želite da napustite pregled?";
+            MessageBoxButton button = MessageBoxButton.OKCancel;
+            MessageBoxResult result = MessageBox.Show(message, "Prekidanje pregleda", button, MessageBoxImage.Warning);
+
+            if (result == MessageBoxResult.Cancel)
+            {
+                return;
+            }
+            else
+            {
+                int thisCount = (this.Parent as Panel).Children.IndexOf(this);
+                (this.Parent as Panel).Children.RemoveRange(3, thisCount);
+                return;
+            }
         }
 
         private void buttonIdiNaDatum_Click(object sender, RoutedEventArgs e)
         {
-            this.Content = new ZakaziKontroluLista();
+            if (calendar.SelectedDate == null)
+            {
+                MessageBox.Show("Morate izabrati jedan datum!", "Greška", MessageBoxButton.OK, MessageBoxImage.Error);
+                return;
+            }
+            DateTime day = (DateTime)calendar.SelectedDate;
+            UserControl usc = new ZakaziKontroluLista(Appointment, day);
+            (this.Parent as Panel).Children.Add(usc);
+        }
+
+        private void backButton_Click(object sender, RoutedEventArgs e)
+        {
+            (this.Parent as Panel).Children.Remove(this);
+            return;
+        }
+
+        private void helpButton_Click(object sender, RoutedEventArgs e)
+        {
+            String message = "Kada izaberete jedan datum i kliknete dugme \"Idi na datum\", dobićete listu zakazanih i slobodnih termina.";
+            MessageBox.Show(message, "Help", MessageBoxButton.OK, MessageBoxImage.Information);
         }
     }
 }
